move utils to export lib
