+++ packages/aws-c2a/lib/index.ts
@@ -1,4 +1,5 @@
-export * from './toolkit';
 export * from './c2a-host';
 export * from './cfn-traverser';
 export * from './cloud-assembly';
+export * from './toolkit';
+export * from './toolkit-utils';
